requirements.txt
streamlit
anthropic
gspread
google-auth
pandas
openpyxl